Cpu optimisation - diminish memory use

+++ GPUPatternMining/GPUPatternMining.CPU/CPUMiningAlgorithms/CPUMiningAlgorithmParallel.h
@@ -20,19 +20,19 @@ public:
 	void loadData(
 		DataFeed* data,
 		size_t size,
-		unsigned int types) override;
+		unsigned short types) override;
 
 	void filterByDistance(float threshold) override;
 	void filterByPrevalence(float prevalence) override;
 	void constructMaximalCliques() override;
-	std::vector<std::vector<unsigned int>> filterMaximalCliques(float prevalence);
+	std::vector<std::vector<unsigned short>> filterMaximalCliques(float prevalence);
 
-	std::map<unsigned int, std::map<unsigned int, std::map<unsigned int, std::vector<unsigned int>*>>> getInsTable()
+	std::map<unsigned short, std::map<unsigned short, std::map<unsigned short, std::vector<unsigned short>*>>> getInsTable()
 	{
 		return insTable;
 	}
 
-	std::vector<std::vector<unsigned int>> getMaximalCliques()
+	std::vector<std::vector<unsigned short>> getMaximalCliques()
 	{
 		return maximalCliques;
 	}
@@ -44,23 +44,23 @@ private:
 	std::vector<int> typeIncidenceCounter;
 	/// InsTable - 2 dimensional hashtable, where frist 2 indexes are types
 	/// the value is a map, where key is number of 1st facility's instanceId and value is a vector of 2nd facility's instancesId 
-	std::map<unsigned int, 
-		std::map<unsigned int,
-			std::map<unsigned int,
-				std::vector<unsigned int>*>>> insTable;
+	std::map<unsigned short, 
+		std::map<unsigned short,
+			std::map<unsigned short,
+				std::vector<unsigned short>*>>> insTable;
 	/// Cm
-	std::vector<std::vector<unsigned int>> maximalCliques;
+	std::vector<std::vector<unsigned short>> maximalCliques;
 	Graph size2ColocationsGraph;
 
-	std::map<std::pair <unsigned int, unsigned int>, std::pair<unsigned int, unsigned int>> countUniqueInstances();
+	std::map<std::pair <unsigned short, unsigned short>, std::pair<unsigned short, unsigned short>> countUniqueInstances();
 
 	bool filterNodeCandidate(
-		unsigned int type,
-		unsigned int instanceId,
+		unsigned short type,
+		unsigned short instanceId,
 		std::vector<CinsNode*> const & ancestors);
 	void createSize2ColocationsGraph();
-	std::vector<std::vector<ColocationElem>> constructCondensedTree(const std::vector<unsigned int>& Cm);
-	bool isCliquePrevalent(std::vector<unsigned int>& clique, float prevalence);
-	std::vector<std::vector<unsigned int>> getPrevalentMaxCliques(std::vector<unsigned int>& clique, float prevalence);
+	std::vector<std::vector<ColocationElem>> constructCondensedTree(const std::vector<unsigned short>& Cm);
+	bool isCliquePrevalent(std::vector<unsigned short>& clique, float prevalence);
+	std::vector<std::vector<unsigned short>> getPrevalentMaxCliques(std::vector<unsigned short>& clique, float prevalence);
 
 };
